Move registration from AssetManager to AssetRegistrar
