+++ src/AssetManager.php
@@ -8,15 +8,8 @@ use RuntimeException;
 use Yiisoft\Aliases\Aliases;
 use Yiisoft\Assets\Exception\InvalidConfigException;
 
-use function array_key_exists;
-use function get_class;
-use function gettype;
 use function in_array;
 use function is_array;
-use function is_file;
-use function is_int;
-use function is_object;
-use function is_string;
 
 /**
  * AssetManager manages asset bundle configuration and loading.
@@ -49,30 +42,9 @@ final class AssetManager
     private array $loadedBundles = [];
     private array $dummyBundles = [];
 
-    /**
-     * @psalm-var CssFile[]
-     */
-    private array $cssFiles = [];
-
-    /**
-     * @psalm-var CssString[]
-     */
-    private array $cssStrings = [];
-
-    /**
-     * @psalm-var JsFile[]
-     */
-    private array $jsFiles = [];
-
-    /**
-     * @psalm-var JsString[]
-     */
-    private array $jsStrings = [];
-
-    private array $jsVars = [];
-    private ?AssetConverterInterface $converter = null;
     private ?AssetPublisherInterface $publisher = null;
     private AssetLoaderInterface $loader;
+    private AssetRegistrar $registrar;
     private Aliases $aliases;
 
     /**
@@ -98,6 +70,7 @@ final class AssetManager
         $this->loader = $loader;
         $this->allowedBundleNames = $allowedBundleNames;
         $this->customizedBundles = $customizedBundles;
+        $this->registrar = new AssetRegistrar($this->aliases, $this->loader);
     }
 
     /**
@@ -149,7 +122,7 @@ final class AssetManager
      */
     public function getCssFiles(): array
     {
-        return $this->cssFiles;
+        return $this->registrar->getCssFiles();
     }
 
     /**
@@ -160,7 +133,7 @@ final class AssetManager
      */
     public function getCssStrings(): array
     {
-        return $this->cssStrings;
+        return $this->registrar->getCssStrings();
     }
 
     /**
@@ -170,7 +143,7 @@ final class AssetManager
      */
     public function getJsFiles(): array
     {
-        return $this->jsFiles;
+        return $this->registrar->getJsFiles();
     }
 
     /**
@@ -181,7 +154,7 @@ final class AssetManager
      */
     public function getJsStrings(): array
     {
-        return $this->jsStrings;
+        return $this->registrar->getJsStrings();
     }
 
     /**
@@ -191,7 +164,7 @@ final class AssetManager
      */
     public function getJsVars(): array
     {
-        return array_values($this->jsVars);
+        return $this->registrar->getJsVars();
     }
 
     /**
@@ -204,7 +177,7 @@ final class AssetManager
     public function withConverter(AssetConverterInterface $converter): self
     {
         $new = clone $this;
-        $new->converter = $converter;
+        $new->registrar = $new->registrar->withConverter($converter);
         return $new;
     }
 
@@ -219,6 +192,7 @@ final class AssetManager
     {
         $new = clone $this;
         $new->loader = $loader;
+        $new->registrar = $new->registrar->withLoader($new->loader);
         return $new;
     }
 
@@ -314,88 +288,6 @@ final class AssetManager
     }
 
     /**
-     * Converter SASS, SCSS, Stylus and other formats to CSS.
-     *
-     * @param AssetBundle $bundle
-     */
-    private function convertCss(AssetBundle $bundle): void
-    {
-        foreach ($bundle->css as $i => $css) {
-            if (is_array($css)) {
-                $file = $css[0];
-                if (AssetUtil::isRelative($file)) {
-                    $baseFile = $this->aliases->get("{$bundle->basePath}/{$file}");
-                    if (is_file($baseFile)) {
-                        /**
-                         * @psalm-suppress PossiblyNullArgument
-                         * @psalm-suppress PossiblyNullReference
-                         */
-                        $css[0] = $this->converter->convert(
-                            $file,
-                            $bundle->basePath,
-                            $bundle->converterOptions,
-                        );
-
-                        $bundle->css[$i] = $css;
-                    }
-                }
-            } elseif (AssetUtil::isRelative($css)) {
-                $baseCss = $this->aliases->get("{$bundle->basePath}/{$css}");
-                if (is_file("$baseCss")) {
-                    /**
-                     * @psalm-suppress PossiblyNullArgument
-                     * @psalm-suppress PossiblyNullReference
-                     */
-                    $bundle->css[$i] = $this->converter->convert(
-                        $css,
-                        $bundle->basePath,
-                        $bundle->converterOptions
-                    );
-                }
-            }
-        }
-    }
-
-    /**
-     * Convert files from TypeScript and other formats into JavaScript.
-     *
-     * @param AssetBundle $bundle
-     */
-    private function convertJs(AssetBundle $bundle): void
-    {
-        foreach ($bundle->js as $i => $js) {
-            if (is_array($js)) {
-                $file = $js[0];
-                if (AssetUtil::isRelative($file)) {
-                    $baseFile = $this->aliases->get("{$bundle->basePath}/{$file}");
-                    if (is_file($baseFile)) {
-                        /**
-                         * @psalm-suppress PossiblyNullArgument
-                         * @psalm-suppress PossiblyNullReference
-                         */
-                        $js[0] = $this->converter->convert(
-                            $file,
-                            $bundle->basePath,
-                            $bundle->converterOptions
-                        );
-
-                        $bundle->js[$i] = $js;
-                    }
-                }
-            } elseif (AssetUtil::isRelative($js)) {
-                $baseJs = $this->aliases->get("{$bundle->basePath}/{$js}");
-                if (is_file($baseJs)) {
-                    /**
-                     * @psalm-suppress PossiblyNullArgument
-                     * @psalm-suppress PossiblyNullReference
-                     */
-                    $bundle->js[$i] = $this->converter->convert($js, $bundle->basePath);
-                }
-            }
-        }
-    }
-
-    /**
      * Registers the named asset bundle.
      *
      * All dependent asset bundles will be registered.
@@ -405,7 +297,7 @@ final class AssetManager
      * This will adjust depending assets javascript file position or fail if requirement can not be met.
      * If this is null, asset bundles position settings will not be changed.
      *
-     * {@see registerJsFile()} For more details on javascript position.
+     * {@see AssetRegistrar::registerJsFile()} For more details on javascript position.
      *
      * @throws InvalidConfigException If the asset or the asset file paths to be published does not exist.
      * @throws RuntimeException If the asset bundle does not exist or a circular dependency is detected.
@@ -474,268 +366,7 @@ final class AssetManager
             $this->registerFiles($dep);
         }
 
-        $this->registerAssetFiles($bundle);
-    }
-
-    /**
-     * Registers asset files from a bundle considering dependencies.
-     *
-     * @param AssetBundle $bundle
-     *
-     * @throws InvalidConfigException If asset files are not found.
-     */
-    private function registerAssetFiles(AssetBundle $bundle): void
-    {
-        if (isset($bundle->basePath, $bundle->baseUrl) && null !== $this->converter) {
-            $this->convertCss($bundle);
-            $this->convertJs($bundle);
-        }
-
-        foreach ($bundle->js as $key => $js) {
-            $this->registerJsFile(
-                $bundle,
-                is_string($key) ? $key : null,
-                $js,
-            );
-        }
-        foreach ($bundle->jsStrings as $key => $jsString) {
-            $this->registerJsString(
-                $bundle,
-                is_string($key) ? $key : null,
-                $jsString,
-            );
-        }
-        foreach ($bundle->jsVars as $name => $jsVar) {
-            if (is_string($name)) {
-                $this->registerJsVar($name, $jsVar, $bundle->jsPosition);
-            } else {
-                $this->registerJsVarByConfig($jsVar, $bundle->jsPosition);
-            }
-        }
-
-        foreach ($bundle->css as $key => $css) {
-            $this->registerCssFile(
-                $bundle,
-                is_string($key) ? $key : null,
-                $css,
-            );
-        }
-        foreach ($bundle->cssStrings as $key => $cssString) {
-            $this->registerCssString(
-                $bundle,
-                is_string($key) ? $key : null,
-                $cssString,
-            );
-        }
-    }
-
-    /**
-     * Registers a CSS file.
-     *
-     * @param array|string $css
-     *
-     * @throws InvalidConfigException
-     */
-    private function registerCssFile(AssetBundle $bundle, ?string $key, $css): void
-    {
-        if (is_array($css)) {
-            if (!array_key_exists(0, $css)) {
-                throw new InvalidConfigException('Do not set in array CSS URL.');
-            }
-            $url = $css[0];
-        } else {
-            $url = $css;
-        }
-
-        if (!is_string($url)) {
-            throw new InvalidConfigException(
-                sprintf(
-                    'CSS file should be string. Got %s.',
-                    $this->getType($url),
-                )
-            );
-        }
-
-        if ($url === '') {
-            throw new InvalidConfigException('CSS file should be non empty string.');
-        }
-
-        $url = $this->loader->getAssetUrl($bundle, $url);
-
-        if (is_array($css)) {
-            $css[0] = $url;
-        } else {
-            $css = [$url];
-        }
-
-        if ($bundle->cssPosition !== null && !isset($css[1])) {
-            $css[1] = $bundle->cssPosition;
-        }
-
-        /** @psalm-var CssFile */
-        $css = $this->mergeOptionsWithArray($bundle->cssOptions, $css);
-
-        $this->cssFiles[$key ?: $url] = $css;
-    }
-
-    /**
-     * Registers a CSS string.
-     *
-     * @param mixed $cssString
-     *
-     * @throws InvalidConfigException
-     */
-    private function registerCssString(AssetBundle $bundle, ?string $key, $cssString): void
-    {
-        if (is_array($cssString)) {
-            $config = $cssString;
-            if (!array_key_exists(0, $config)) {
-                throw new InvalidConfigException('CSS string do not set in array.');
-            }
-        } else {
-            $config = [$cssString];
-        }
-
-        if ($bundle->cssPosition !== null && !isset($config[1])) {
-            $config[1] = $bundle->cssPosition;
-        }
-
-        /** @psalm-var CssString */
-        $config = $this->mergeOptionsWithArray($bundle->cssOptions, $config);
-
-        if ($key === null) {
-            $this->cssStrings[] = $config;
-        } else {
-            $this->cssStrings[$key] = $config;
-        }
-    }
-
-    /**
-     * Registers a JS file.
-     *
-     * @param array|string $js
-     *
-     * @throws InvalidConfigException
-     */
-    private function registerJsFile(AssetBundle $bundle, ?string $key, $js): void
-    {
-        if (is_array($js)) {
-            if (!array_key_exists(0, $js)) {
-                throw new InvalidConfigException('Do not set in array JS URL.');
-            }
-            $url = $js[0];
-        } else {
-            $url = $js;
-        }
-
-        if (!is_string($url)) {
-            throw new InvalidConfigException(
-                sprintf(
-                    'JS file should be string. Got %s.',
-                    $this->getType($url),
-                )
-            );
-        }
-
-        if ($url === '') {
-            throw new InvalidConfigException('JS file should be non empty string.');
-        }
-
-        $url = $this->loader->getAssetUrl($bundle, $url);
-
-        if (is_array($js)) {
-            $js[0] = $url;
-        } else {
-            $js = [$url];
-        }
-
-        if ($bundle->jsPosition !== null && !isset($js[1])) {
-            $js[1] = $bundle->jsPosition;
-        }
-
-        /** @psalm-var JsFile */
-        $js = $this->mergeOptionsWithArray($bundle->jsOptions, $js);
-
-        $this->jsFiles[$key ?: $url] = $js;
-    }
-
-    /**
-     * Registers a JS string.
-     *
-     * @param array|string $jsString
-     *
-     * @throws InvalidConfigException
-     */
-    private function registerJsString(AssetBundle $bundle, ?string $key, $jsString): void
-    {
-        if (is_array($jsString)) {
-            if (!array_key_exists(0, $jsString)) {
-                throw new InvalidConfigException('JavaScript string do not set in array.');
-            }
-        } else {
-            $jsString = [$jsString];
-        }
-
-        if ($bundle->jsPosition !== null && !isset($jsString[1])) {
-            $jsString[1] = $bundle->jsPosition;
-        }
-
-        /** @psalm-var JsString */
-        $jsString = $this->mergeOptionsWithArray($bundle->jsOptions, $jsString);
-
-        if ($key === null) {
-            $this->jsStrings[] = $jsString;
-        } else {
-            $this->jsStrings[$key] = $jsString;
-        }
-    }
-
-    /**
-     * Registers a JavaScript variable.
-     *
-     * @param mixed $value
-     */
-    private function registerJsVar(string $name, $value, ?int $position): void
-    {
-        $config = [$name, $value];
-
-        if ($position !== null) {
-            $config[2] = $position;
-        }
-
-        $this->jsVars[$name] = $config;
-    }
-
-    /**
-     * Registers a JavaScript variable by config.
-     *
-     * @throws InvalidConfigException
-     */
-    private function registerJsVarByConfig(array $config, ?int $bundleJsPosition): void
-    {
-        if (!array_key_exists(0, $config)) {
-            throw new InvalidConfigException('Do not set JavaScript variable name.');
-        }
-        $name = $config[0];
-
-        if (!is_string($name)) {
-            throw new InvalidConfigException(
-                sprintf(
-                    'JavaScript variable name should be string. Got %s.',
-                    $this->getType($name),
-                )
-            );
-        }
-
-        if (!array_key_exists(1, $config)) {
-            throw new InvalidConfigException('Do not set JavaScript variable value.');
-        }
-        /** @var mixed */
-        $value = $config[1];
-
-        $position = $config[2] ?? $bundleJsPosition;
-
-        $this->registerJsVar($name, $value, $position);
+        $this->registrar->register($bundle);
     }
 
     /**
@@ -832,31 +463,5 @@ final class AssetManager
         }
 
         return false;
-    }
-
-    /**
-     * @throws InvalidConfigException
-     */
-    private function mergeOptionsWithArray(array $options, array $array): array
-    {
-        foreach ($options as $key => $value) {
-            if (is_int($key)) {
-                throw new InvalidConfigException(
-                    'JavaScript or CSS options should be list of key/value pairs with string keys. Got integer key.'
-                );
-            }
-            if (!array_key_exists($key, $array)) {
-                $array[$key] = $value;
-            }
-        }
-        return $array;
-    }
-
-    /**
-     * @param mixed $value
-     */
-    private function getType($value): string
-    {
-        return is_object($value) ? get_class($value) : gettype($value);
     }
 }
